Integrate dynamic HDHG utilities into augmentations
