+++ src/models/gnn/__init__.py
@@ -1,3 +1,8 @@
 from .graphcl_encoder import GraphCLEncoder
 from .denoising_gcn import DenoisingGCN
-from .augmentations import drop_nodes, permute_edges
+from .augmentations import (
+    drop_nodes,
+    permute_edges,
+    DynamicHypergraphPartitioner,
+    HDHGEncoder,
+)
