--- src/models/gnn/dynamic_hdhg.py
@@ -1,338 +1,0 @@
-# src/models/gnn/dynamic_hdhg.py
-"""
-Dynamic HDHG utilities
-~~~~~~~~~~~~~~~~~~~~~~
-
-- DynamicHypergraphPartitioner
-    * build()        : 최초 하이퍼그래프 구성
-    * step()         : (선택) 주기적 재구성
-- HDHGEncoder
-    * forward(x)     : HypergraphConv ▶ 비선형 ▶ 투영
-- Example usage at bottom
-
-The design follows:
-(1) Local k-NN hyperedges for fine-grained proximity   [Jiang+2019, DHGNN]
-(2) Global k-means clusters for coarse communities     [Shi+2000, Zhou+2006]
-(3) Multi-channel heterogeneous partitions per rel-type
-    with learnable fusion weights                      [Zhao+2024, Li+2023]
-(4) Optional prune / grow hooks (HSL-style)            [Cai+2022]
-"""
-
-from __future__ import annotations
-from typing import Dict, List, Optional, Tuple
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch_geometric.nn import HypergraphConv
-from sklearn.cluster import KMeans
-from torch_geometric.utils import to_dense_adj
-
-
-class DynamicHypergraphPartitioner:
-    """
-    Build & refresh hyperedges on demand.
-
-    Parameters
-    ----------
-    mode : {"knn", "cluster", "hetero"}
-        'knn'      – each node + k nearest neighbours → 1 hyperedge
-        'cluster'  – k-means on embeddings → cluster-wise hyperedge
-        'hetero'   – split by relation type; each channel holds its own
-                     partition strategy; see `hetero_cfg`
-    k : int, default=5
-        Neighbourhood size for mode='knn'
-    num_clusters : int, default=10
-        k-means cluster count for mode='cluster'
-    hetero_cfg : dict[str, dict]
-        For mode='hetero'. Key = relation/channel name.
-        Value = {"mode": "knn"/"cluster", "k": 3, "num_clusters": 8, "weight": 1.0}
-    prune_ratio : float, default=0.0
-        Fraction of *least-important* hyperedges to drop each refresh step
-        (§3.2, Cai+2022 HSL) – importance≔mean edge weight.
-    device : torch.device
-    """
-
-    def __init__(
-        self,
-        mode: str = "knn",
-        k: int = 5,
-        num_clusters: int = 10,
-        hetero_cfg: Optional[Dict[str, Dict]] = None,
-        prune_ratio: float = 0.0,
-        device: torch.device | str = "cpu",
-    ):
-        self.mode = mode
-        self.k = k
-        self.num_clusters = num_clusters
-        self.hetero_cfg = hetero_cfg or {}
-        self.prune_ratio = prune_ratio
-        self.device = device
-
-        # Filled after build()
-        self.hyperedge_index: Optional[torch.Tensor] = None
-        self.channel_ptr: Optional[List[Tuple[int, int]]] = None  # (start, end)
-        self.weights: Optional[torch.Tensor] = None               # edge weights
-
-    # --------------------------------------------------------------------- #
-    # Public API
-    # --------------------------------------------------------------------- #
-    def build(
-        self,
-        X: torch.Tensor,
-        rel_labels: Optional[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        """
-        Forward-only build (no gradients).
-
-        Parameters
-        ----------
-        X : (N, F) float tensor – node features / embeddings
-        rel_labels : (N,) int tensor – node-type or relation-type id
-                      (needed when mode='hetero')
-
-        Returns
-        -------
-        hyperedge_index : (2, E) long tensor
-            Row-0: node ids ; Row-1: hyperedge ids.
-        """
-        with torch.no_grad():
-            if self.mode == "knn":
-                he = self._knn_hyperedges(X, self.k)
-            elif self.mode == "cluster":
-                he = self._cluster_hyperedges(X, self.num_clusters)
-            elif self.mode == "hetero":
-                if rel_labels is None:
-                    raise ValueError("rel_labels required for hetero mode")
-                he = self._hetero_hyperedges(X, rel_labels)
-            else:
-                raise ValueError(f"Unsupported mode: {self.mode}")
-
-            he = self._prune(he, self.prune_ratio)
-            self.hyperedge_index, self.channel_ptr, self.weights = he
-            return self.hyperedge_index
-
-    def step(
-        self,
-        X: torch.Tensor,
-        epoch: int,
-        refresh_every: int = 5,
-        rel_labels: Optional[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        """
-        Optionally rebuild hyperedges every `refresh_every` epochs (HSL idea).
-
-        Returns current `hyperedge_index` (unchanged if no refresh this step).
-        """
-        if epoch % refresh_every == 0:
-            return self.build(X, rel_labels)
-        return self.hyperedge_index
-
-    # --------------------------------------------------------------------- #
-    # Internal helpers
-    # --------------------------------------------------------------------- #
-    def _knn_hyperedges(
-        self, X: torch.Tensor, k: int
-    ) -> Tuple[torch.Tensor, List[Tuple[int, int]], torch.Tensor]:
-        N = X.size(0)
-        dist = (X.square().sum(1, keepdim=True) -
-                2 * X @ X.t() +
-                X.square().sum(1, keepdim=True).t())
-        _, nn_idx = dist.topk(k + 1, largest=False)  # self + k-nn
-        nn_idx = nn_idx[:, 1:]                       # drop self
-
-        edges_i, edges_j, w = [], [], []
-        eid = 0
-        for node in range(N):
-            members = [node] + nn_idx[node].tolist()
-            edges_i.extend(members)
-            edges_j.extend([eid] * len(members))
-            w.extend([1.0] * len(members))
-            eid += 1
-        return (torch.tensor([edges_i, edges_j], dtype=torch.long, device=self.device),
-                [(0, eid)], torch.tensor(w, device=self.device))
-
-    def _cluster_hyperedges(
-        self, X: torch.Tensor, C: int
-    ) -> Tuple[torch.Tensor, List[Tuple[int, int]], torch.Tensor]:
-        X_np = X.cpu().numpy()
-        labels = KMeans(n_clusters=C, n_init=10).fit_predict(X_np)
-        edges_i, edges_j, w = [], [], []
-        for eid in range(C):
-            members = (labels == eid).nonzero()[0]
-            edges_i.extend(members)
-            edges_j.extend([eid] * len(members))
-            w.extend([1.0] * len(members))
-        return (torch.tensor([edges_i, edges_j], dtype=torch.long, device=self.device),
-                [(0, len(set(labels)))],
-                torch.tensor(w, device=self.device))
-
-    def _hetero_hyperedges(
-        self,
-        X: torch.Tensor,
-        rel_labels: torch.Tensor,
-    ) -> Tuple[torch.Tensor, List[Tuple[int, int]], torch.Tensor]:
-        edges_i, edges_j, w = [], [], []
-        channel_ptr = []
-        eid_cursor = 0
-
-        for rtype, cfg in self.hetero_cfg.items():
-            mask = rel_labels == int(rtype)
-            if mask.sum() == 0:
-                continue
-            X_sub = X[mask]
-            idx_map = mask.nonzero(as_tuple=True)[0]
-            if cfg["mode"] == "knn":
-                he, _, _ = self._knn_hyperedges(X_sub, cfg.get("k", self.k))
-            elif cfg["mode"] == "cluster":
-                he, _, _ = self._cluster_hyperedges(
-                    X_sub, cfg.get("num_clusters", self.num_clusters)
-                )
-            else:
-                raise ValueError(f"Unknown mode in hetero_cfg[{rtype}]")
-
-            # Remap local node ids → global ids
-            he_remap = he.clone()
-            he_remap[0] = idx_map[he[0]]
-            he_remap[1] += eid_cursor
-
-            edges_i.extend(he_remap[0].tolist())
-            edges_j.extend(he_remap[1].tolist())
-            w.extend([cfg.get("weight", 1.0)] * he_remap.size(1))
-
-            channel_ptr.append((eid_cursor, he_remap[1].max().item() + 1))
-            eid_cursor = channel_ptr[-1][1]
-
-        return (torch.tensor([edges_i, edges_j], dtype=torch.long, device=self.device),
-                channel_ptr,
-                torch.tensor(w, device=self.device))
-
-    def _prune(
-        self,
-        he: Tuple[torch.Tensor, List[Tuple[int, int]], torch.Tensor],
-        ratio: float
-    ) -> Tuple[torch.Tensor, List[Tuple[int, int]], torch.Tensor]:
-        if ratio <= 0:
-            return he
-        hyperedge_index, channel_ptr, weights = he
-        E = hyperedge_index[1].max().item() + 1
-        he_sizes = torch.bincount(hyperedge_index[1], minlength=E).float()
-        # Importance ≔ size * mean weight (simple proxy)
-        importance = he_sizes / he_sizes.max()
-        prune_k = int(E * ratio)
-        if prune_k == 0:
-            return he
-        keep_mask = torch.ones(E, dtype=torch.bool, device=self.device)
-        keep_mask[importance.topk(prune_k, largest=False).indices] = False
-
-        # Filter edges
-        edge_keep = keep_mask[hyperedge_index[1]]
-        hyperedge_index = hyperedge_index[:, edge_keep]
-        weights = weights[edge_keep]
-
-        # Re-index hyperedge ids to 0…E'-1
-        unique_old = hyperedge_index[1].unique(sorted=True)
-        remap = torch.full((E,), -1, dtype=torch.long, device=self.device)
-        remap[unique_old] = torch.arange(unique_old.size(0), device=self.device)
-        hyperedge_index[1] = remap[hyperedge_index[1]]
-        # Channel_ptr also remap
-        # (simple: rebuild from scratch per contiguous blocks)
-        new_ptr = []
-        start = 0
-        for _, ch_end in channel_ptr:
-            kept = keep_mask[_:ch_end]
-            cnt = kept.sum().item()
-            if cnt > 0:
-                new_ptr.append((start, start + cnt))
-                start += cnt
-        channel_ptr = new_ptr
-        return hyperedge_index, channel_ptr, weights
-
-
-class HDHGEncoder(nn.Module):
-    """
-    Simple 2-layer HypergraphConv encoder with channel-wise attention fusion.
-
-    Parameters
-    ----------
-    in_dim : int
-    hid_dim : int
-    out_dim : int
-    num_heads : int
-        Attention heads for channel fusion (≥ number of channels).
-    dropout : float
-    """
-    def __init__(
-        self,
-        in_dim: int,
-        hid_dim: int = 64,
-        out_dim: int = 64,
-        num_heads: int = 4,
-        dropout: float = 0.1,
-    ):
-        super().__init__()
-        self.hgc1 = HypergraphConv(in_dim, hid_dim)
-        self.hgc2 = HypergraphConv(hid_dim, out_dim)
-        self.dropout = dropout
-        self.act = nn.ReLU()
-
-        self.attn = nn.MultiheadAttention(out_dim, num_heads, dropout=dropout, batch_first=True)
-
-    def forward(
-        self,
-        X: torch.Tensor,
-        hyperedge_index: torch.Tensor,
-        channel_ptr: Optional[List[Tuple[int, int]]] = None,
-    ) -> torch.Tensor:
-        # HypergraphConv layers
-        h = self.act(self.hgc1(X, hyperedge_index))
-        h = F.dropout(h, p=self.dropout, training=self.training)
-        h = self.hgc2(h, hyperedge_index)  # shape (N, out_dim)
-
-        if channel_ptr:  # Channel-wise fusion
-            channel_embeds = []
-            for start, end in channel_ptr:
-                # Build dense adj for this channel (for pooling)
-                he_idx = hyperedge_index.clone()
-                ch_mask = (he_idx[1] >= start) & (he_idx[1] < end)
-                he_idx = he_idx[:, ch_mask]
-                adj = to_dense_adj(he_idx).squeeze(0)  # (N, N)
-                if adj.sum() == 0:
-                    channel_embeds.append(h)
-                else:
-                    pooled = adj @ h / adj.sum(dim=-1, keepdim=True).clamp(min=1e-6)
-                    channel_embeds.append(pooled)
-            h = torch.stack(channel_embeds, dim=1)  # (N, C, out_dim)
-            h, _ = self.attn(h, h, h)  # self-attn fuse channels
-            h = h.mean(dim=1)          # aggregate heads/channels
-        return h
-
-
-# ------------------------------------------------------------------------- #
-# Minimal example
-# ------------------------------------------------------------------------- #
-if __name__ == "__main__":
-    torch.manual_seed(0)
-    N, F = 50, 16
-    X = torch.randn(N, F)
-
-    # Suppose we have 3 relation types (0/1/2) per node
-    rel_labels = torch.randint(0, 3, (N,))
-
-    hetero_cfg = {
-        "0": {"mode": "knn", "k": 3, "weight": 1.0},
-        "1": {"mode": "cluster", "num_clusters": 4, "weight": 1.2},
-        "2": {"mode": "knn", "k": 5, "weight": 0.8},
-    }
-
-    builder = DynamicHypergraphPartitioner(
-        mode="hetero",
-        hetero_cfg=hetero_cfg,
-        prune_ratio=0.1,
-        device="cpu",
-    )
-    he_index = builder.build(X, rel_labels)
-    model = HDHGEncoder(in_dim=F, hid_dim=32, out_dim=32, num_heads=4)
-
-    z = model(X, he_index, builder.channel_ptr)
-    print("Node embeddings:", z.shape)  # -> (N, 32)
